Optimized Value Parsing in Card.hpp
Now calls to value() will not parse the value again and again. Rather they are parsed once and cached inside the variant. Whenever programmer requests the value the cache_value is returned. If the value was changed by the programmer, the cache will automatically become invalid and the new value will be parsed and cached for optimized future calls